Return CompletableFuture instead of Publisher in HttpHandler
HttpHandler#handle() return value purpose is to indicate when the handler
execution is complete, or to return an exception, both asynchronously.
CompletableFuture<Void> is exactly designed for this use case, not need
of Publisher additional behaviors (subscription,  back-pressure).

This change also remove the need for unneeded Publisher -> Reactor Stream
wrapping.